cleanup
cleaning up spaghetti is always a good thing.

Took out the remaining unneeded casts, which would just add lag and do absolutely nothing in the end. You're welcome, future me.

+++ src/main/java/com/vetpetmon/wyrmsofnyrus/creativetab/TabWyrms.java
@@ -22,7 +22,7 @@ public class TabWyrms extends AutoReg.ModElement {
 			@SideOnly(Side.CLIENT)
 			@Override
 			public ItemStack getTabIconItem() {
-				return new ItemStack(ItemWyrmico.block, (int) (1));
+				return new ItemStack(ItemWyrmico.block, 1);
 			}
 
 			@SideOnly(Side.CLIENT)
